requirements.txt
llama-index==0.12.49
openai
streamlit
python-dotenv
mysql-connector-python
matplotlib
pandas
tqdm
